refactor: one repo-key derivation behind the issues and notes stores
Both stores keyed their records by `git rev-parse --git-common-dir` and
named the repo by the first `git worktree list` entry, in private copies
that had drifted three ways: the empty-worktree-list case (toplevel
fallback vs throw), the not-a-git-repo error (raw vs mapped), and whether
a non-empty input was checked at all.

`repo-key.ts` owns the derivation; `gitMainWorktree` answers null for
"no worktree line" so each store's own `resolveRepo` keeps deciding what
that means. No behaviour changes — both policies survive verbatim.
